classificador-lixo/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
scikit-learn==1.4.2
scikit-image==0.23.2
pillow==10.3.0
numpy==1.26.4
joblib==1.4.2
python-multipart==0.0.9
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
